// src/app/interfaces/mis/datadirectory.ts
export interface IDataDirectory {
  id: number
  name: string
  contactPerson: string
  typeId: number
  dataTypeName: string
  pincode: string
  address: string
  contactNo: string
  emailid: string
  website: string
  longitude: string
  latitude: string
  stateId: number
  stateName: string
  cityId: number
  cityName: string
  countryId: number
  countryName: string
  facebook: string
  youTube: string
  twitter: string
  instagram: string
}